Python Flask/App-Demo/requirements.txt
astroid==2.2.5
beautifulsoup4==4.8.0
Click==7.0
colorama==0.4.1
Flask==1.1.1
isort==4.3.21
itsdangerous==1.1.0
Jinja2==2.10.1
lazy-object-proxy==1.4.2
lxml==4.4.1
MarkupSafe==1.1.1
mccabe==0.6.1
Pillow==6.1.0
pylint==2.3.1
python-dotenv==0.10.3
redis==3.3.8
rq==1.1.0
six==1.12.0
soupsieve==1.9.4
typed-ast==1.4.0
Werkzeug==0.15.6
wrapt==1.11.2
